Add new 'Service' section to application
This commit introduces a new 'Service' section in both the 'exo' and 'demo' modules of the application. The new features include the creation of several new components such as 'InstructionsComponent', 'DemoServicesComponent', and 'ExoServiceComponent'. It also includes the implementation of services ('ListeService' and 'ScoreService') as well as updates to the application routing to accommodate these changes.