Implement user login/logout functionality
This commit introduces a user login/logout feature by adding 'LoginService', 'User' model, and 'Login' component. It also includes the addition of a new guard 'GEOFFRE_ONLY' in the routing module which allows access only to user Geoffery. This functionality enables the users to log in with a username and password, and log out when needed.